requirements.txt
Django
django-bootstrap3
django-debug-toolbar
django-environ
django-extensions
